requirements.txt
asarPy==1.0.1
mysql-connector-python==9.6.0
